+++ src/Application.kt
@@ -1,7 +1,6 @@
 package team.uptech.food.bot
 
 import com.google.gson.Gson
-import com.google.gson.annotations.SerializedName
 import io.ktor.application.Application
 import io.ktor.application.call
 import io.ktor.application.log
@@ -20,27 +19,28 @@ import kotlinx.coroutines.withContext
 import team.uptech.food.bot.bot.Bot
 import team.uptech.food.bot.bot.BotReplay
 import team.uptech.food.bot.data.Storage
+import team.uptech.food.bot.data.respones.Payload
 import team.uptech.food.bot.di.Injector.injectApplication
-import team.uptech.food.bot.di.Injector.injectSubComponentNewOrder
-import team.uptech.food.bot.presentation.new_order.NewOrderPresenter
+import team.uptech.food.bot.presentation.Context
+import team.uptech.food.bot.presentation.new_order.makeNewOrder
+import team.uptech.food.bot.presentation.new_order.processNewOrder
 import team.uptech.food.bot.utils.extensions.configure
 import javax.inject.Inject
 
+
 fun main(args: Array<String>): Unit = io.ktor.server.netty.EngineMain.main(args)
+
 
 class EntryPoint(application: Application) {
 
   @Inject
   lateinit var storage: Storage
+
   @Inject
   lateinit var botReplay: BotReplay
 
-  @Inject
-  lateinit var newOrderPresenter: NewOrderPresenter
-
   init {
     injectApplication(application)
-    injectSubComponentNewOrder()?.inject(this)
   }
 }
 
@@ -51,12 +51,20 @@ fun Application.module(testing: Boolean = false) {
 
   val client = HttpClient(Apache)
   val entry = EntryPoint(this)
+  val context = Context(this, client)
 
   routing {
-    get(Bot.HOME_PATH) { call.respondText(entry.botReplay.hello(), contentType = ContentType.Text.Plain) }
+
+    get(Bot.HOME_PATH) {
+      call.respondText(entry.botReplay.hello(), contentType = ContentType.Text.Plain)
+    }
 
     post(Bot.NEW_ORDER_PATH) {
-      entry.newOrderPresenter.startNewOrder(this)
+      context.call = call
+
+      withContext(Dispatchers.Default) {
+        makeNewOrder(context)
+      }
     }
 
     post(Bot.MANUAL_RESET_PATH) {
@@ -74,27 +82,24 @@ fun Application.module(testing: Boolean = false) {
     }
 
     post(Bot.USER_INTERACTIONS) {
+      context.call = call
+
       withContext(Dispatchers.Default) {
         val parameters = call.receiveParameters()
         log.debug(parameters.entries().toString())
 
+        val payload = Gson().fromJson(parameters[Payload.KEY], Payload::class.java)
+
         // FIXME: hardcoded parameters 👎🏻
-
-        val view = Gson().fromJson(parameters["payload"], Payload::class.java)
-
-        if (view?.view != null && "order_initiation" == view.view.callbackId) {
-          entry.newOrderPresenter.processNewOrder(this@post, parameters["payload"])
+        // FIXME: Use block_id and action_id instead of callback_id
+        if (payload?.view != null && "order_initiation" == payload.view.callbackId) {
+          processNewOrder(context, payload)
         } else {
           call.respond(HttpStatusCode.OK, "")
         }
       }
 
-      // TODO: add fields validation for presentation.modals
+      // TODO: add fields validation for presentation.dsl.modals
     }
   }
 }
-
-/* TODO: extract from here to data layer */
-data class Payload(@SerializedName("view") val view: View)
-
-data class View(@SerializedName("callback_id") val callbackId: String)
