+++ src/data/respones/User.kt
@@ -1,0 +1,6 @@
+package team.uptech.food.bot.data.respones
+
+import com.google.gson.annotations.SerializedName
+
+
+data class User(@SerializedName("id") val id: String)
