+++ src/di/Injector.kt
@@ -3,12 +3,10 @@ package team.uptech.food.bot.di
 import io.ktor.application.Application
 import team.uptech.food.bot.di.components.AppComponent
 import team.uptech.food.bot.di.components.DaggerAppComponent
-import team.uptech.food.bot.di.components.NewOrderComponent
 import team.uptech.food.bot.di.modules.ApplicationModule
 
 object Injector {
   private var applicationComponent: AppComponent? = null
-  private var newOrderComponent: NewOrderComponent? = null
 
   fun injectApplication(application: Application): AppComponent? =
     if (applicationComponent != null) {
@@ -23,17 +21,4 @@ object Injector {
   fun dropAppComponent() {
     applicationComponent = null
   }
-
-  fun injectSubComponentNewOrder() =
-    if (newOrderComponent != null) {
-      newOrderComponent
-    } else {
-      applicationComponent?.plusNewOrderComponent()
-        .also { newOrderComponent = it }
-    }
-
-  fun dropSubComponentNewOrder() {
-    newOrderComponent = null
-  }
-
 }
