+++ src/domain/order/model/Order.kt
@@ -2,7 +2,6 @@ package team.uptech.food.bot.domain.order.model
 
 import team.uptech.food.bot.domain.client.model.Client
 import team.uptech.food.bot.domain.initiator.model.Initiator
-import team.uptech.food.bot.slack.interactions.Channel
 
 data class Order(
   val id: Long,
