+++ src/presentation/dsl/errors/ErrorBuilder.kt
@@ -1,0 +1,10 @@
+package team.uptech.food.bot.presentation.dsl.errors
+
+object ErrorBuilder {
+  fun assembleNewOrderError() = NewOrderError()
+    .assemble {
+      errors {
+        "" to ""
+      }
+    }
+}
