+++ src/presentation/dsl/models/BuilderMarker.kt
@@ -1,0 +1,4 @@
+package team.uptech.food.bot.presentation.dsl.models
+
+@DslMarker
+annotation class BuilderMarker
