--- src/presentation/errors/ErrorBuilder.kt
@@ -1,10 +1,0 @@
-package team.uptech.food.bot.presentation.errors
-
-object ErrorBuilder {
-  fun assembleNewOrderError() = NewOrderError()
-    .assemble {
-      errors {
-        "" to ""
-      }
-    }
-}
