--- src/presentation/errors/NewOrderError.kt
@@ -1,8 +1,0 @@
-package team.uptech.food.bot.presentation.errors
-
-import team.uptech.food.bot.presentation.errors.models.Error
-
-class NewOrderError : BaseError() {
-  override fun assemble(block: Error.() -> Unit): String = NewOrderError().apply(
-    Error().apply(block))
-}
