--- src/presentation/messages/NewOrderSettingsMessage.kt
@@ -1,7 +1,0 @@
-package team.uptech.food.bot.presentation.messages
-
-import team.uptech.food.bot.presentation.messages.models.Message
-
-class NewOrderSettingsMessage : BaseMessage() {
-  override fun assemble(block: Message.() -> Unit): String = NewOrderSettingsMessage().apply(Message().apply(block))
-}
